Move CreateStore
